refactor - moving redux things around
