--- hooks/useAppDispatch.ts
@@ -1,7 +1,0 @@
-import { AppDispatch } from "@/store/store";
-import { useDispatch } from "react-redux";
-
-export default function useAppDispatch() {
-    const dispatch = useDispatch<AppDispatch>();
-    return dispatch;
-}
