+++ redux/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { TypedUseSelectorHook, useSelector } from "react-redux";
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 
 import authReducer from "./features/auth-slice"
 
@@ -15,3 +15,4 @@ export type AppDispatch = typeof store.dispatch;
 
 // export custom useSelector hook
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+export const dispatch = useDispatch<AppDispatch>();
